requirements.txt
streamlit
streamlit_extras
streamlit-drawable-canvas
pymixbox
colormath
colorspacious
watchdog